디버깅

+++ app/src/main/java/com/example/dbvideomarker/activity/PlayListEditActivity.java
@@ -6,14 +6,12 @@ import android.os.Bundle;
 import android.preference.PreferenceManager;
 import android.view.Menu;
 import android.view.MenuInflater;
-import android.view.MenuItem;
 import android.view.View;
 import android.widget.PopupMenu;
 import android.widget.TextView;
 
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.ItemTouchHelper;
@@ -25,12 +23,10 @@ import com.bumptech.glide.RequestManager;
 import com.example.dbvideomarker.R;
 import com.example.dbvideomarker.adapter.PlayList_MarkAdapter;
 import com.example.dbvideomarker.adapter.PlayList_VideoAdapter;
-import com.example.dbvideomarker.database.entitiy.PlRelMark;
 import com.example.dbvideomarker.database.entitiy.PlRelVideo;
 import com.example.dbvideomarker.listener.OnItemClickListener;
 import com.example.dbvideomarker.adapter.util.Callback;
 import com.example.dbvideomarker.database.entitiy.PlRel;
-import com.example.dbvideomarker.database.entitiy.PlayList;
 import com.example.dbvideomarker.listener.OnMarkClickListener;
 import com.example.dbvideomarker.player.PlayerActivity;
 import com.example.dbvideomarker.ui.home.HomeViewModel;
@@ -56,7 +52,8 @@ public class PlayListEditActivity extends AppCompatActivity implements OnItemCli
 
     private ItemTouchHelper itemTouchHelper;
     private RequestManager _mGlideRequestManager;
-    private FloatingActionButton fab_main, fab_video, fab_mark;
+    private FloatingActionButton fab_video;
+    private FloatingActionButton fab_mark;
     private TextView vCount, mCount;
     private PlayList_VideoAdapter adapter_video;
     private PlayList_MarkAdapter adapter_mark;
@@ -84,74 +81,56 @@ public class PlayListEditActivity extends AppCompatActivity implements OnItemCli
         playlistViewModel = new ViewModelProvider(this).get(PlaylistViewModel.class);
         homeViewModel = new ViewModelProvider(this).get(HomeViewModel.class);
         playListEditViewModel = new ViewModelProvider(this).get(PlayListEditViewModel.class);
-        playListEditViewModel.getPlayList(pid).observe(this, new Observer<PlayList>() {
-            @Override
-            public void onChanged(PlayList playList) {
-                String pname = playList.getpName();
-                playListName.setText(pname);
-            }
+        playListEditViewModel.getPlayList(pid).observe(this, playList -> {
+            String pname = playList.getpName();
+            playListName.setText(pname);
         });
 
         vCount = findViewById(R.id.video_count);
         mCount = findViewById(R.id.mark_count);
 
-        playListEditViewModel.getVideoRowCount(pid).observe(this, new Observer<Integer>() {
-            @Override
-            public void onChanged(Integer integer) {
-                VIDEO_COUNT = integer;
-                vCount.setText(String.valueOf(integer));
-            }
+        playListEditViewModel.getVideoRowCount(pid).observe(this, integer -> {
+            VIDEO_COUNT = integer;
+            vCount.setText(String.valueOf(integer));
         });
-        playListEditViewModel.getMarkRowCount(pid).observe(this, new Observer<Integer>() {
-            @Override
-            public void onChanged(Integer integers) {
-                MARK_COUNT = integers;
-                mCount.setText(String.valueOf(integers));
-            }
+        playListEditViewModel.getMarkRowCount(pid).observe(this, integers -> {
+            MARK_COUNT = integers;
+            mCount.setText(String.valueOf(integers));
         });
         setVideoInPlaylist();
         setMarkInPlaylist();
         setFab();
     }
 
-    public void setFab(){
-        fab_main = findViewById(R.id.fab_main);
+    public void setFab() {
+        FloatingActionButton fab_main = findViewById(R.id.fab_main);
         fab_video = findViewById(R.id.fab_video);
         fab_mark = findViewById(R.id.fab_mark);
 
-        fab_main.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                if(!IS_OPEN) {
-                    fab_video.setVisibility(View.VISIBLE);
-                    fab_mark.setVisibility(View.VISIBLE);
-                    IS_OPEN = true;
-                } else {
-                    fab_video.setVisibility(View.GONE);
-                    fab_mark.setVisibility(View.GONE);
-                    IS_OPEN = false;
-                }
+        fab_main.setOnClickListener(view -> {
+            if (!IS_OPEN) {
+                fab_video.setVisibility(View.VISIBLE);
+                fab_mark.setVisibility(View.VISIBLE);
+                IS_OPEN = true;
+            } else {
+                fab_video.setVisibility(View.GONE);
+                fab_mark.setVisibility(View.GONE);
+                IS_OPEN = false;
             }
         });
 
-        fab_video.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent videoIntent = new Intent(PlayListEditActivity.this, SelectActivity.class);
-                videoIntent.putExtra("pid", pid);
-                videoIntent.putExtra("VIEW_TYPE", 2001);
-                startActivityForResult(videoIntent, SELECT_VIDEO_REQUEST_CODE);
-            }
+        fab_video.setOnClickListener(view -> {
+            Intent videoIntent = new Intent(PlayListEditActivity.this, SelectActivity.class);
+            videoIntent.putExtra("pid", pid);
+            videoIntent.putExtra("VIEW_TYPE", 2001);
+            startActivityForResult(videoIntent, SELECT_VIDEO_REQUEST_CODE);
         });
 
-        fab_mark.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent markIntent = new Intent(PlayListEditActivity.this, SelectActivity.class);
-                markIntent.putExtra("pid", pid);
-                markIntent.putExtra("VIEW_TYPE", 2002);
-                startActivityForResult(markIntent, SELECT_MARK_REQUEST_CODE);
-            }
+        fab_mark.setOnClickListener(view -> {
+            Intent markIntent = new Intent(PlayListEditActivity.this, SelectActivity.class);
+            markIntent.putExtra("pid", pid);
+            markIntent.putExtra("VIEW_TYPE", 2002);
+            startActivityForResult(markIntent, SELECT_MARK_REQUEST_CODE);
         });
     }
 
@@ -165,41 +144,38 @@ public class PlayListEditActivity extends AppCompatActivity implements OnItemCli
 
         //vCount = findViewById(R.id.video_count);
 
-        playListEditViewModel.findVideoInPlayList(pid).observe(this, new Observer<List<PlRelVideo>>() {
-            @Override
-            public void onChanged(List<PlRelVideo> plRels) {
-                adapter_video.setPlRelv(plRels);
-                //vCount.setText(""+plRels.size());
+        playListEditViewModel.findVideoInPlayList(pid).observe(this, plRels -> {
+            adapter_video.setPlRelv(plRels);
+            //vCount.setText(""+plRels.size());
 /*
-                resultList = getStringArrayList(""+pid);
+            resultList = getStringArrayList(""+pid);
 
-                if(resultList == null) {
-                    setStringArrayList(""+pid, plRels);
-                    adapter.setPlRels(plRels);
-                } else {
-                    if(resultList.size() < plRels.size()) {
-                        for(int i = 0; i < plRels.size(); i++) {
-                            if (!resultList.contains(plRels.get(i))) {
-                                resultList.add(plRels.get(i));
-                                setStringArrayList(""+pid, resultList);
-                                adapter.setPlRels(resultList);
-                            }
+            if(resultList == null) {
+                setStringArrayList(""+pid, plRels);
+                adapter.setPlRels(plRels);
+            } else {
+                if(resultList.size() < plRels.size()) {
+                    for(int i = 0; i < plRels.size(); i++) {
+                        if (!resultList.contains(plRels.get(i))) {
+                            resultList.add(plRels.get(i));
+                            setStringArrayList(""+pid, resultList);
+                            adapter.setPlRels(resultList);
                         }
-                    } else if(resultList.size() > plRels.size()){
-                        for(int i = 0; i < resultList.size(); i++) {
-                            if (!plRels.contains(resultList.get(i))) {
-                                resultList.remove(i);
-                                setStringArrayList(""+pid, resultList);
-                                adapter.setPlRels(resultList);
-                            }
-                        }
-                    } else {
-                        setStringArrayList(""+pid, resultList);
-                        adapter.setPlRels(resultList);
                     }
+                } else if(resultList.size() > plRels.size()){
+                    for(int i = 0; i < resultList.size(); i++) {
+                        if (!plRels.contains(resultList.get(i))) {
+                            resultList.remove(i);
+                            setStringArrayList(""+pid, resultList);
+                            adapter.setPlRels(resultList);
+                        }
+                    }
+                } else {
+                    setStringArrayList(""+pid, resultList);
+                    adapter.setPlRels(resultList);
                 }
- */
             }
+*/
         });
 
         recyclerView_video.setLayoutManager(new LinearLayoutManager(this, LinearLayoutManager.VERTICAL, false));
@@ -213,12 +189,9 @@ public class PlayListEditActivity extends AppCompatActivity implements OnItemCli
 
         //mCount = findViewById(R.id.mark_count);
 
-        playListEditViewModel.findMarkInPlayList(pid).observe(this, new Observer<List<PlRelMark>>() {
-            @Override
-            public void onChanged(List<PlRelMark> plRelMarks) {
-                adapter_mark.setPlRelm(plRelMarks);
-                //mCount.setText(""+plRelMarks.size());
-            }
+        playListEditViewModel.findMarkInPlayList(pid).observe(this, plRelMarks -> {
+            adapter_mark.setPlRelm(plRelMarks);
+            //mCount.setText(""+plRelMarks.size());
         });
         recyclerView_mark.setLayoutManager(new GridLayoutManager(this, 2));
         recyclerView_mark.setAdapter(adapter_mark);
@@ -228,21 +201,25 @@ public class PlayListEditActivity extends AppCompatActivity implements OnItemCli
     protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
 
-        if(resultCode == RESULT_OK) {
-            if(requestCode == SELECT_VIDEO_REQUEST_CODE) {
+        if (resultCode == RESULT_OK) {
+            if (requestCode == SELECT_VIDEO_REQUEST_CODE) {
+                assert data != null;
                 ArrayList<Integer> selectedVidList = data.getIntegerArrayListExtra("vidlist");
                 playlistViewModel.updateVideoCount(pid, VIDEO_COUNT);
-                for(int i=0; i<selectedVidList.size(); i++) {
+                assert selectedVidList != null;
+                for (int i = 0; i < selectedVidList.size(); i++) {
                     PlRel plRel = new PlRel();
                     plRel.setPid(pid);
                     plRel.setMid(-1);
                     plRel.setVid(selectedVidList.get(i));
                     playListEditViewModel.insertPlRelation(plRel);
                 }
-            } else if(requestCode == SELECT_MARK_REQUEST_CODE) {
+            } else if (requestCode == SELECT_MARK_REQUEST_CODE) {
+                assert data != null;
                 ArrayList<Integer> selectedMidList = data.getIntegerArrayListExtra("midlist");
                 playlistViewModel.updateMarkCount(pid, MARK_COUNT);
-                for(int i=0; i<selectedMidList.size(); i++) {
+                assert selectedMidList != null;
+                for (int i = 0; i < selectedMidList.size(); i++) {
                     PlRel plRel = new PlRel();
                     plRel.setPid(pid);
                     plRel.setMid(selectedMidList.get(i));
@@ -269,16 +246,11 @@ public class PlayListEditActivity extends AppCompatActivity implements OnItemCli
         Menu menu = popupMenu.getMenu();
         inflater.inflate(R.menu.menu_popup_playlist_video, menu);
 
-        popupMenu.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
-            @Override
-            public boolean onMenuItemClick(MenuItem menuItem) {
-                switch (menuItem.getItemId()) {
-                    case(R.id.popup_delete_playlist):
-                        playListEditViewModel.deleteVideoInPlaylist(id);
-                        break;
-                }
-                return false;
+        popupMenu.setOnMenuItemClickListener(menuItem -> {
+            if (menuItem.getItemId() == R.id.popup_delete_playlist) {
+                playListEditViewModel.deleteVideoInPlaylist(id);
             }
+            return false;
         });
         popupMenu.show();
     }
@@ -300,16 +272,11 @@ public class PlayListEditActivity extends AppCompatActivity implements OnItemCli
         Menu menu = popupMenu.getMenu();
         inflater.inflate(R.menu.menu_popup_playlist_mark, menu);
 
-        popupMenu.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
-            @Override
-            public boolean onMenuItemClick(MenuItem menuItem) {
-                switch (menuItem.getItemId()) {
-                    case(R.id.popup_delete_playlist):
-                        playListEditViewModel.deleteMarkInPlaylist(id);
-                        break;
-                }
-                return false;
+        popupMenu.setOnMenuItemClickListener(menuItem -> {
+            if (menuItem.getItemId() == R.id.popup_delete_playlist) {
+                playListEditViewModel.deleteMarkInPlaylist(id);
             }
+            return false;
         });
         popupMenu.show();
     }
@@ -322,13 +289,13 @@ public class PlayListEditActivity extends AppCompatActivity implements OnItemCli
     @Override
     protected void onPause() {
         super.onPause();
-        setStringArrayList(""+pid, adapter_video.plRelList);
+        setStringArrayList("" + pid, adapter_video.plRelList);
     }
 
     @Override
     protected void onDestroy() {
         super.onDestroy();
-        setStringArrayList(""+pid, adapter_video.plRelList);
+        setStringArrayList("" + pid, adapter_video.plRelList);
     }
 
     public void setStringArrayList(String key, List<PlRelVideo> valueList) {
@@ -338,7 +305,7 @@ public class PlayListEditActivity extends AppCompatActivity implements OnItemCli
         Gson gson = new Gson();
         String json = gson.toJson(valueList);
         editor.putString(key, json);
-        editor.commit();
+        editor.apply();
     }
 
     public List<PlRelVideo> getStringArrayList(String key) {
@@ -347,7 +314,6 @@ public class PlayListEditActivity extends AppCompatActivity implements OnItemCli
         String json = pref.getString(key, null);
         Type type = new TypeToken<List<PlRelVideo>>() {
         }.getType();
-        List<PlRelVideo> items = gson.fromJson(json, type);
-        return items;
+        return gson.fromJson(json, type);
     }
 }
